Create e-commerce website.html
a simple e commerce webpage used to make a efficient website which helps users to interact with different products with the help of html, css, and javascript in frontend and php and my sql in backend.  proficiently used with the both of the technologies and helps in better user response.